meminfo lib - partially revert 4264873ab27931b6848e67e6e103632ac1abdb7b
From a mail by Jaromir Capik <jcapik@redhat.com> in the procps mailing list):

... few people were strictly against the subtraction and called the
change as reinforcing the misconception that memory in the page cache
can be considered as free.

The kernel guys explained, that the Cached value is not a sum of page
cache and tmpfs, as in fact the tmpfs memory lives in the page cache
(= is a part of) and therefore shouldn't be subtracted.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>