check_memory / meminfo lib - rework the used/cached/free memory stats
meminfo library: the memory usage calculation has been reviewed and is
now more accurate. The changes are based on the (long) discussions and
patches that have been recently posted in the procps mailing list.

check_memory: the command line options '-C|--caches' and '-a|--available'
now do nothing but are accepted for compatibility.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>