docker: query docker daemon via the socket file
The sysfs exported metrics are not always available.
Switch to queries of the docker daemon via its socket file.
This implementation requires libcurl development files.

Note that the plugin check_docker does not currently return
a correct output. A json parser needs to be added in order
to process the output provided by docker.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>