build: fix the libprocps detection at build time
Most distro packaged version of procps-ng correctly report the
version of the library libprocps.

According to the project mailing-list and my own tests, Debian
and Ubuntu report UNKNOWN as library version.
But, under openSUSE Tumbleweed, Archlinux, Slackware, and Fedora
it is shown properly.

So switch from "UNKNOWN" to "libprocps >= 3.3.12" in configure.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>