check_network: exit with an unknown state if -% cannot be used
If the percentage output is selected along with a warning and/or
threshold and an interface matching the interfaces regexp is down
or its physical speed is not available for this type of interface,
do make the plugin exit with an unknown state.
Otherwise the thresholds will be used for both percentages and
counters, which does not make sense.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>