check_network: make it possible to set thresholds for all metrics
Add the plugins:

 * check_network_collisions
 * check_network_dropped
 * check_network_errors
 * check_network_multicast

to make it possible to set the warning and critical thresholds for
all the (interesting) metrics and not only for traffic in bytes,
as it was previously the case.

Add itwo extra command-line switches '-r/--rx-only' and '-t/--tx-only'
for discarding the transmission and received metrics respectively.
This should allow even more custom plugin configurations and should be
especially usefull for setting thresholds in the network traffic
in bytes.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>